Service volume tests

+++ test-all.sh
@@ -84,6 +84,9 @@ testScratchPush () {
 
 
 runTests() {
+  export X_TEST_SERVICE_VOL_PATH=$testsDir/test-service-vol
+  basicTest "service volume" build "$testsDir/service-volume" --enable-volumes || return 1
+  grep -q "test-volume-file" "${workingDir}/service volume.log" || return 1
   basicTest "source-path" build "$testsDir/source-path" || return 1
   basicTest "rm pipeline" build --artifacts "$testsDir/rm-pipeline" || return 1
   basicTest "local services" build "$testsDir/local-service/service-consumer" || return 1
